Split POVTurret into ManualTurret and ManualHood
