+++ src/main/java/frc/robot/commands/hood/ManualHood.java
@@ -1,0 +1,53 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.commands.hood;
+
+import java.util.function.DoubleSupplier;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.Hood;
+
+public class ManualHood extends CommandBase {
+
+  Hood m_hood;
+  DoubleSupplier m_speedSupplier;
+
+  /**
+   * Creates a new ManualHood.
+   */
+  public ManualHood(DoubleSupplier speedSupplier, Hood hood) {
+    m_hood = hood;
+    m_speedSupplier = speedSupplier;
+    addRequirements(m_hood);
+  }
+
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {
+    m_hood.disable();
+  }
+
+  // Called every time the scheduler runs while the command is scheduled.
+  @Override
+  public void execute() {
+    m_hood.set(m_speedSupplier.getAsDouble());
+  }
+
+  // Called once the command ends or is interrupted.
+  @Override
+  public void end(boolean interrupted) {
+    m_hood.stop();
+    m_hood.enable();
+  }
+
+  // Returns true when the command should end.
+  @Override
+  public boolean isFinished() {
+    return false;
+  }
+}
